+++ clj/src/limabean.clj
@@ -1,61 +1,26 @@
 (ns limabean
-  "Top-level limabean functions for use from the REPL."
   (:require [clojure.java.io :as io]
-            [limabean.adapter.loader :as loader]
-            [limabean.adapter.logging :as logging]
-            [limabean.adapter.show :as show]
-            [limabean.adapter.pod :as pod]
-            [limabean.adapter.bean-queries :as bean-queries]))
+            [limabean.adapter.json]
+            [limabean.adapter.print]
+            [limabean.core.filters :as f]
+            [limabean.core.inventory :as inventory]
+            [limabean.core.xf :as xf]
+            [limabean.core.journal :as journal]
+            [limabean.core.registry :as registry]
+            [limabean.core.rollup :as rollup]
+            [limabean.adapter.loader :as loader]))
 
-(def ^:dynamic *beans*
-  "An aggregate of the elements which were used in deriving the directives for the current beanfile.
 
-  Useful in case of failed plugins, for inspecting partial state."
-  nil)
-(def ^:dynamic *directives*
-  "Vector of all directives form the beanfile after running plugins."
-  nil)
-(def ^:dynamic *options* "Map of options from the beanfile." nil)
-(def ^:dynamic *registry*
-  "Map of attributes derived from directives and options, e.g. booking method for account."
-  nil)
-
-(defn- assign-limabean-globals
-  [beans]
-  (let [directives (get beans :directives [])
-        options (get beans :options {})]
-    (alter-var-root #'*beans* (constantly beans))
-    (alter-var-root #'*directives* (constantly directives))
-    (alter-var-root #'*options* (constantly options))
-    (alter-var-root #'*registry* (constantly (:registry beans)))))
-
-(defn load-beanfile
-  [path]
-  (when (:pod *beans*) (pod/stop (:pod *beans*)))
-  (assign-limabean-globals {})
-  (logging/initialize)
-  (let [beans (loader/load-beanfile path)]
-    (binding [*out* *err*]
-      (println "[limabean]" (count (:raw-directives beans))
-               "directives loaded from" path)
-      (let [bad-plugins (filter :err (:plugins beans))]
-        (doseq [plugin bad-plugins]
-          (println "ERROR in plugin" (:name plugin)
-                   "-" (get-in plugin [:err :message]))))
-      (assign-limabean-globals beans)
-      (if-let [err (:plugin-errors beans)]
-        (println err)
-        (println "[limabean]"
-                 (count (:directives beans))
-                 "directives resulting from booking and running plugins")))
-    :ok))
+(defn- postings
+  [directives filters]
+  (eduction (comp (xf/postings) (xf/all-of filters)) directives))
 
 (defn inventory
-  "Build inventory from `*beans*` after applying filters, if any.
-
-  Custom directives may be passed in after the filters using :directives."
-  [& args]
-  (bean-queries/inventory *beans* args))
+  "Build inventory from `beans` after applying filters, if any."
+  ([beans] (inventory beans []))
+  ([beans filters]
+   (inventory/build (postings (:directives beans) filters)
+                    (partial registry/acc-booking (:registry beans)))))
 
 (defn rollup
   "Build a rollup for the primary currency from an inventory.
@@ -66,32 +31,41 @@
   ```
   "
   [inv]
-  (bean-queries/rollup inv))
+  (let [primary-cur (first (apply max-key val (inventory/cur-freq inv)))]
+    (rollup/build inv primary-cur)))
 
 (defn balances
-  "Build balances from `*beans*`, optionally further filtered.
-
-  Custom directives may be passed in after the filters using :directives.
-  "
-  [& args]
-  (bean-queries/balances *beans* args))
+  "Build balances from `beans`, optionally further filtered."
+  ([beans] (balances beans []))
+  ([beans filters]
+   (inventory beans
+              (conj filters
+                    (f/sub-acc (:name-assets (:options beans))
+                               (:name-liabilities (:options beans)))))))
 
 (defn income-statement
-  "Build balances from `*beans*`, optionally further filtered.
+  "Build balances from `beans`, optionally further filtered.
 
   Custom directives may be passed in after the filters using :directives.
   "
-  [& args]
-  (bean-queries/income-statement *beans* args))
+  ([beans] (income-statement beans []))
+  ([beans filters]
+   (inventory beans
+              (conj filters
+                    (f/sub-acc (:name-income (:options beans))
+                               (:name-expenses (:options beans)))))))
 
 (defn journal
-  "Build a journal of postings from `*beans*` with running balance.
+  "Build a journal of postings from `beans` with running balance.
 
-  Custom directives may be passed in after the filters using :directives."
-  [& args]
-  (bean-queries/journal *beans* args))
+   Custom directives may be passed in after the filters using :directives."
+  ([beans] (journal beans []))
+  ([beans filters] (journal/build (postings (:directives beans) filters))))
 
-(defn show "Convert `x` to a cell and tabulate it." [x] (show/show x))
+(defn load-beanfile
+  "Load beans from the beanfile at path"
+  [path]
+  (loader/load-beanfile path))
 
 (defn version
   "Get the library version from pom.properties, else returns \"unknown\"."